// packages/by-admin-api/src/core/config.enum.ts
export enum ConfigEnum {
  NODE_ENV = "NODE_ENV",
  // App 相关配置
  APP_PORT = "APP_PORT",
  // MySQL 相关配置
  MYSQL_HOST = "MYSQL_HOST",
  MYSQL_PORT = "MYSQL_PORT",
  MYSQL_DATABASE = "MYSQL_DATABASE",
  MYSQL_USERNAME = "MYSQL_USERNAME",
  MYSQL_PASSWORD = "MYSQL_PASSWORD",

  // Redis 相关配置
  REDIS_HOST = "REDIS_HOST",
  REDIS_PORT = "REDIS_PORT",
  REDIS_DB = "REDIS_DB",
  REDIS_PASSWORD = "REDIS_PASSWORD",
  LOG_ON = "LOG_ON",
  LOG_LEVEL = "LOG_LEVEL"
}